+++ Editor/DecalDataEditor.cs
@@ -1,11 +1,12 @@
 ﻿using UnityEngine;
 using UnityEditor;
+using UnityEditorInternal;
 
 namespace kTools.Decals.Editor
 {
     using Editor = UnityEditor.Editor;
 
-    [CustomEditor(typeof(DecalData))]
+    [CustomEditor(typeof(DecalData)), CanEditMultipleObjects]
     sealed class DecalDataEditor : Editor
     {
 #region Structs
@@ -23,14 +24,34 @@ namespace kTools.Decals.Editor
                 "How many instances of this Decal will be created in the Pool.");
 
             public static readonly GUIContent Depth = new GUIContent("Depth",
-                "How far the decal projection should blend on Z axis.");
+                "How far the Decal projection draw on Z axis.");
+
+            public static readonly GUIContent DepthFalloff = new GUIContent("Depth Falloff",
+                "How much the Decal should blend transparency towards its Depth value.");
+
+            public static readonly GUIContent Angle = new GUIContent("Angle",
+                "Maximum angle between surface and Decal forward vector.");
+
+            public static readonly GUIContent AngleFalloff = new GUIContent("Angle Falloff",
+                "How much the Decal should blend transparency towards its Angle value.");
+
+            public static readonly GUIContent LayerMask = new GUIContent("Layer Mask",
+                "Which layers the Decal is applied to.");
+
+            public static readonly GUIContent SortingOrder = new GUIContent("Sorting Order",
+                "Decals with higher values are drawn on top of ones with lower values.");
         }
 
         struct PropertyNames
         {
             public static readonly string PoolingEnabled = "m_PoolingEnabled";
             public static readonly string InstanceCount = "m_InstanceCount";
-            public static readonly string ProjectionDepth = "m_ProjectionDepth";
+            public static readonly string Depth = "m_Depth";
+            public static readonly string DepthFalloff = "m_DepthFalloff";
+            public static readonly string Angle = "m_Angle";
+            public static readonly string AngleFalloff = "m_AngleFalloff";
+            public static readonly string LayerMask = "m_LayerMask";
+            public static readonly string SortingOrder = "m_SortingOrder";
         }
 #endregion
 
@@ -46,7 +67,12 @@ namespace kTools.Decals.Editor
         // Properties
         SerializedProperty m_PoolingEnabledProp;
         SerializedProperty m_InstanceCountProp;
-        SerializedProperty m_ProjectionDepthProp;
+        SerializedProperty m_DepthProp;
+        SerializedProperty m_DepthFalloffProp;
+        SerializedProperty m_AngleProp;
+        SerializedProperty m_AngleFalloffProp;
+        SerializedProperty m_LayerMaskProp;
+        SerializedProperty m_SortingOrderProp;
 #endregion
 
 #region State
@@ -60,7 +86,12 @@ namespace kTools.Decals.Editor
             // Get Properties
             m_PoolingEnabledProp = serializedObject.FindProperty(PropertyNames.PoolingEnabled);
             m_InstanceCountProp = serializedObject.FindProperty(PropertyNames.InstanceCount);
-            m_ProjectionDepthProp = serializedObject.FindProperty(PropertyNames.ProjectionDepth);
+            m_DepthProp = serializedObject.FindProperty(PropertyNames.Depth);
+            m_DepthFalloffProp = serializedObject.FindProperty(PropertyNames.DepthFalloff);
+            m_AngleProp = serializedObject.FindProperty(PropertyNames.Angle);
+            m_AngleFalloffProp = serializedObject.FindProperty(PropertyNames.AngleFalloff);
+            m_LayerMaskProp = serializedObject.FindProperty(PropertyNames.LayerMask);
+            m_SortingOrderProp = serializedObject.FindProperty(PropertyNames.SortingOrder);
 
             // Create Editors
             m_MaterialEditor = Editor.CreateEditor(m_Target.material) as MaterialEditor;
@@ -134,12 +165,36 @@ namespace kTools.Decals.Editor
 
         void DrawProjectionOptions()
         {
-            // Projection Depth
-            EditorGUILayout.PropertyField(m_ProjectionDepthProp, Styles.Depth);
+            // Depth
+            EditorGUILayout.PropertyField(m_DepthProp, Styles.Depth);
+            EditorGUILayout.Slider(m_DepthFalloffProp, 0.0f, 1.0f, Styles.DepthFalloff);
+
+            // Angle
+            EditorGUILayout.Slider(m_AngleProp, 0.0f, 180.0f, Styles.Angle);
+            EditorGUILayout.Slider(m_AngleFalloffProp, 0.0f, 1.0f, Styles.AngleFalloff);
+
+            // Layer Mask
+            EditorGUI.BeginChangeCheck();
+            LayerMask tempMask = EditorGUILayout.MaskField(Styles.LayerMask, (LayerMask)m_LayerMaskProp.intValue, InternalEditorUtility.layers);
+            if(EditorGUI.EndChangeCheck())
+            {
+                m_LayerMaskProp.intValue = (int)tempMask;
+            }
+
+            // Sorting Order
+            EditorGUILayout.PropertyField(m_SortingOrderProp, Styles.SortingOrder);
         }
 
         void DrawMaterialEditor()
         {
+            // Multiple selection cant be handled for Material inspector...
+            var selectionCount = Selection.objects.Length;
+            if(selectionCount > 1)
+            {
+                EditorGUILayout.HelpBox("Multiple Object editing is not supported for Material Options.", MessageType.Warning);
+                return;
+            }
+
             // Default Material Editor
             EditorGUILayout.Space();
             m_MaterialEditor.DrawHeader();