+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this package are documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - 2020-01-20
+## [2.0.0] - 2020-01-24
 ### Added
 - `BaseGUI` for custom Decal shaders
 - Full Summary for all public API
+- Angle and angle falloff controls
+- Layer masking
+- Sorting order
 
 ### Changed
 - Moved to package
@@ -23,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Expose projection depth to `DecalData`
 - Improve shader GUI
 - Serialize foldouts to EditorPrefs
+- Converted to Universal render pipeline
+- Removed requirements for `Projector` component
 
 ## [1.0.0] - 2018-11-17
 ### Added